% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_deprecated.R
\name{make_original_db}
\alias{make_original_db}
\title{Armado de bases para RSL (deprecated)}
\arguments{
\item{file_path}{Ruta donde se encuentran los archivos de la búsqueda.}

\item{file_name}{Nombre de los archivos de la búsqueda.}

\item{search_num}{Número de la búsqueda.}
}
\value{
Invisible TRUE.
}
\description{
Armado de bases para RSL (deprecated)
}
